requirements.txt
uvicorn
fastapi
langchain_openai
langchain
langchain_groq
pydantic_settings
langchain_community